refactor(radio): reorganize Radio module structure; rename imports and files; remove obsolete genres sample data
