--- src/modules/Content/Radio/sample-data/genres-data.ts
@@ -1,59 +1,0 @@
-// Sample data for the Genre collection with subgenres of Techno
-const genres = [
-  {
-    id: 1,
-    title: 'Peak Time / Driving',
-    slug: 'peak-time-driving',
-    description:
-      'A high-energy subgenre of techno with relentless beats and driving rhythms, often played during peak hours.',
-    Tracks: {
-      docs: [1, 4, 6, 10], // Tracks: "Aurora", "Eclipse", "Stratosphere", "Momentum"
-      hasNextPage: false,
-    },
-  },
-  {
-    id: 2,
-    title: 'Melodic Techno',
-    slug: 'melodic-techno',
-    description:
-      'A subgenre characterized by emotive melodies layered over hypnotic beats and atmospheric soundscapes.',
-    Tracks: {
-      docs: [3, 9], // Tracks: "Afterlife", "Dreamstate"
-      hasNextPage: false,
-    },
-  },
-  {
-    id: 3,
-    title: 'Dark Techno',
-    slug: 'dark-techno',
-    description:
-      'A moody and intense subgenre of techno, often featuring darker tonalities and powerful, driving rhythms.',
-    Tracks: {
-      docs: [2, 8], // Tracks: "Neural Collapse", "Reverie"
-      hasNextPage: false,
-    },
-  },
-  {
-    id: 4,
-    title: 'Raw / Hypnotic',
-    slug: 'raw-hypnotic',
-    description:
-      'A stripped-down, repetitive style of techno with hypnotic elements and a raw edge.',
-    Tracks: {
-      docs: [5], // Track: "Pulse"
-      hasNextPage: false,
-    },
-  },
-  {
-    id: 5,
-    title: 'Live Techno Sets',
-    slug: 'live-techno-sets',
-    description: 'Live recorded techno sets showcasing extended mixes and unique performances.',
-    Tracks: {
-      docs: [7], // Track: "Genesis"
-      hasNextPage: false,
-    },
-  },
-]
-
-export default genres
