+++ src/modules/Settings/Help/components/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Banner } from '@payloadcms/ui/elements/Banner'
 import { SeedButton } from '@/components/BeforeDashboard/SeedButton'
 import classes from './index.module.scss'
-import { SeedRadioButton } from '@/modules/Content/Radio/seed/SeedRadioButton'
+import { SeedRadioButton } from '@/modules/Radio/seed/SeedRadioButton'
 
 export const HelpComponent: React.FC = () => {
   return (
