+++ src/payload.config.ts
@@ -9,9 +9,9 @@ import { Media } from './modules/Upload/Media'
 import { Pages } from './modules/Content/Pages'
 import { Posts } from './modules/Content/Posts'
 import { Users } from './modules/Settings/Users'
-import { Tracks } from './modules/Content/Radio/Tracks'
-import { Artists } from './modules/Content/Radio/Artists'
-import Genres from './modules/Content/Radio/Genres'
+import { Tracks } from './modules/Radio/Tracks'
+import { Artists } from './modules/Radio/Artists'
+import { Genres } from './modules/Radio/Genres'
 import { Footer } from './modules/Navigation/Footer/config'
 import { Header } from './modules/Navigation/Header/config'
 import { plugins } from './plugins'
