+++ src/plugins/seo.ts
